finish product sv

+++ src/modules/products/products.controller.ts
@@ -3,11 +3,16 @@ import {
   Controller,
   Delete,
   Get,
+  HttpException,
+  HttpStatus,
   Param,
   Patch,
   Post,
+  ValidationPipe,
 } from '@nestjs/common';
 import { ProductsService } from './products.service';
+import CreateProductDto from './dto/create-product.dto';
+import UpdateProductDto from './dto/update-product.dto';
 
 @Controller('products')
 export class ProductsController {
@@ -16,17 +21,30 @@ export class ProductsController {
   findAll() {
     return this.productsService.findAll();
   }
+
+  @Get(':id')
+  async findOne(@Param('id') id: number) {
+    const product = await this.productsService.findOne(id);
+    if (!product) {
+      throw new HttpException('Product not found', HttpStatus.NOT_FOUND);
+    }
+    return product;
+  }
   @Post()
-  create(@Body() productData: any) {
+  create(@Body(new ValidationPipe()) productData: CreateProductDto) {
     return this.productsService.create(productData);
   }
   @Patch(':id')
-  update(@Param('id') id: number, @Body() ProductData: any) {
+  update(@Param('id') id: number, @Body() ProductData: UpdateProductDto) {
     return this.productsService.update(id, ProductData);
   }
 
   @Delete(':id')
-  delete(@Param('id') id: number) {
+  async delete(@Param('id') id: number) {
+    const product = await this.productsService.findOne(id);
+    if (!product) {
+      throw new HttpException('Product not found', HttpStatus.NOT_FOUND);
+    }
     return this.productsService.delete(id);
   }
 }
